refactor : ChartManager로 따로 빼기
