+++ presentation/src/main/java/com/example/presentation/ui/analysis/TotalAnalysisFragment.kt
@@ -1,22 +1,14 @@
 package com.example.presentation.ui.analysis
 
 import android.annotation.SuppressLint
-import android.graphics.Color
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import com.example.presentation.R
 import com.example.presentation.databinding.FragmentTotalAnalysisBinding
-import com.github.mikephil.charting.animation.Easing
-import com.github.mikephil.charting.charts.LineChart
-import com.github.mikephil.charting.components.Description
-import com.github.mikephil.charting.components.XAxis
-import com.github.mikephil.charting.components.YAxis
 import com.github.mikephil.charting.data.*
-import com.github.mikephil.charting.utils.ColorTemplate
 
 
 class TotalAnalysisFragment : Fragment() {
@@ -37,6 +29,12 @@ class TotalAnalysisFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        val keywordEntries: ArrayList<PieEntry> = ArrayList()
+        keywordEntries.add(PieEntry(34f, "NETWORK"))
+        keywordEntries.add(PieEntry(23f, "ALGORITHM"))
+        keywordEntries.add(PieEntry(10f, "JAVA"))
+        keywordEntries.add(PieEntry(3f, "ANDROID"))
+
         val eyesEntries: ArrayList<Entry> = ArrayList()
         eyesEntries.add(Entry(1f, 5f))
         eyesEntries.add(Entry(2f, 8f))
@@ -56,107 +54,27 @@ class TotalAnalysisFragment : Fragment() {
         poseEntries.add(Entry(7f, 8f))
 
         setNotChangeSeekbar()
-        setPieChartKeyword()
-        setLineChart(eyesEntries, binding.lineChartEyes)
-        setLineChart(poseEntries, binding.lineChartPose)
+        ChartManager.setPieChartKeyword(
+            keywordEntries,
+            binding.pieChartKeyword,
+            getString(R.string.percent_of_keyword)
+        )
+        ChartManager.setLineChart(
+            eyesEntries,
+            binding.lineChartEyes,
+            getString(R.string.eyes_contact_by_turns)
+        )
+        ChartManager.setLineChart(
+            poseEntries,
+            binding.lineChartPose,
+            getString(R.string.pose_analysis_by_turns)
+        )
     }
 
     @SuppressLint("ClickableViewAccessibility")
     private fun setNotChangeSeekbar() {
         binding.seekBar.setOnTouchListener { _, _ ->
             true
-        }
-    }
-
-    private fun setPieChartKeyword() {
-        binding.pieChartKeyword.apply {
-            setUsePercentValues(true)
-            description.isEnabled = false
-            setExtraOffsets(5f, 10f, 5f, 5f)
-            isDrawHoleEnabled = true
-            setHoleColor(Color.WHITE)
-            transparentCircleRadius = 61f
-
-            val yValues: ArrayList<PieEntry> = ArrayList()
-            with(yValues) {
-                add(PieEntry(34f, "NETWORK"))
-                add(PieEntry(23f, "ALGORITHM"))
-                add(PieEntry(10f, "JAVA"))
-                add(PieEntry(3f, "ANDROID"))
-            }
-            animateY(1000, Easing.EaseInOutCubic)
-
-            val dataSet: PieDataSet = PieDataSet(yValues, "")
-            with(dataSet) {
-                sliceSpace = 5f
-                selectionShift = 5f
-                setColors(*ColorTemplate.JOYFUL_COLORS)
-            }
-
-            val pieData = PieData(dataSet)
-            with(pieData) {
-                setValueTextSize(15f)
-                setValueTextColor(Color.BLACK)
-                setEntryLabelColor(Color.BLACK)
-                val des = Description()
-                des.text = "답변한 키워드 비율"
-                des.textColor = Color.BLACK
-                des.textSize = 13f
-                description = des
-            }
-
-            data = pieData
-        }
-    }
-
-    private fun setLineChart(entries: ArrayList<Entry>, lineChart: LineChart) {
-        lineChart.apply {
-            val lineDataSet = LineDataSet(entries, "회차별 시선 처리")
-            lineDataSet.setLineWidth(2f)
-            lineDataSet.setCircleRadius(6f)
-            lineDataSet.setCircleColor(
-                ContextCompat.getColor(
-                context,
-                R.color.deep_blue
-            ))
-            lineDataSet.circleHoleColor = ContextCompat.getColor(
-                context,
-                R.color.deep_blue
-            )
-            lineDataSet.setColor(ContextCompat.getColor(
-                context,
-                R.color.sky_blue
-            ))
-            lineDataSet.setDrawCircleHole(true)
-            lineDataSet.setDrawCircles(true)
-            lineDataSet.setDrawHorizontalHighlightIndicator(false)
-            lineDataSet.setDrawHighlightIndicators(false)
-            lineDataSet.setDrawValues(false)
-
-            val lineData = LineData(lineDataSet)
-            setData(lineData)
-
-            val xAxis: XAxis = getXAxis()
-            xAxis.position = XAxis.XAxisPosition.BOTTOM
-            xAxis.textColor = Color.BLACK
-            xAxis.enableGridDashedLine(8f, 24f, 0f)
-
-            val yLAxis: YAxis = getAxisLeft()
-            yLAxis.textColor = Color.BLACK
-
-            val yRAxis: YAxis = getAxisRight()
-            yRAxis.setDrawLabels(false)
-            yRAxis.setDrawAxisLine(false)
-            yRAxis.setDrawGridLines(false)
-
-            val description = Description()
-            description.text = ""
-
-            setDoubleTapToZoomEnabled(false)
-            setDrawGridBackground(false)
-            setDescription(description)
-            animateY(2000, Easing.EaseInCubic)
-            invalidate()
         }
     }
 }
