+++ presentation/src/main/res/values/strings.xml
@@ -71,4 +71,7 @@
     <string name="keyword">키워드</string>
     <string name="eyes_contact">시선 처리</string>
     <string name="pose_analysis">자세 분석</string>
+    <string name="eyes_contact_by_turns">회차별 시선 처리</string>
+    <string name="pose_analysis_by_turns">회차별 자세 분석</string>
+    <string name="percent_of_keyword">답변한 키워드 비율</string>
 </resources>
